% Generated by roxygen2 (4.0.0): do not edit by hand
\name{as.char}
\alias{as.char}
\title{converts to character with minimal loss of precision for numeric variables}
\usage{
as.char(x, ...)
}
\arguments{
\item{x}{the value}

\item{...}{passed to either format or as.character.}
}
\description{
converts to character with minimal loss of precision for numeric variables
}

