% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.R
\name{calculate_no_common_trees}
\alias{calculate_no_common_trees}
\title{calculate_no_common_trees}
\usage{
calculate_no_common_trees(k, m)
}
\arguments{
\item{k}{number of remaining samples}

\item{m}{number of metastases samples}
}
\value{
number of monophyletic trees
}
\description{
Calculate the number of monophyletic trees given by Theorem 3 in the SI of Reiter et al. 2020 Nat Gen.
}
