% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.R
\name{calculate_no_trees}
\alias{calculate_no_trees}
\title{calculate_no_trees}
\usage{
calculate_no_trees(n)
}
\arguments{
\item{n}{number of samples}
}
\value{
number of distinct trees
}
\description{
Calculate the total number of distinct trees given by Edwards and Cavalli-Sforza, 1963
}
