% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.R
\name{calculate_rds}
\alias{calculate_rds}
\title{calculate_rds}
\usage{
calculate_rds(k, m, l, use_gmp)
}
\arguments{
\item{k}{number of primary tumor region samples}

\item{m}{total number of metastases samples}

\item{l}{number of metastases that actually cluster together}

\item{use_gmp}{Use gmp R package to calculate RDS for large trees. Required for trees with more around 30 tips (default=F)}
}
\value{
probabilility that a cluster of size l arises by chance
}
\description{
Calculates the root diversity scores as the probability that a tree where l metastases form a cluster (have one root) exists by chance as defined by Equation (2).
}
