% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.interval.estimate.R
\name{export.rds.interval.estimate}
\alias{export.rds.interval.estimate}
\title{Convert the output of print.rds.interval.estimate from a character data.frame to a numeric matrix}
\usage{
export.rds.interval.estimate(x, proportion = TRUE)
}
\arguments{
\item{x}{An object, typically the result of print.rds.interval.estimate.}

\item{proportion}{logical, Should the outcome be treated as a proportion and converted to a percentage.}
}
\description{
Convert the output of print.rds.interval.estimate from a character data.frame to a numeric matrix
}
