% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds-doc.R
\docType{data}
\name{fauxtime}
\alias{fauxtime}
\title{A Simulated RDS Data Set}
\format{
An rds.data.frame object
}
\description{
This is a faux set used to demonstrate RDS functions and analysis.
}
\references{
Gile, Krista J., Handcock, Mark S., 2010 \emph{Respondent-driven Sampling: An Assessment of Current Methodology},  \emph{Sociological Methodology}, 40, 285-327.
}
\seealso{
\code{\link{fauxsycamore}}, \code{\link{fauxmadrona}}
}
\keyword{datasets}
