% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.data.frame-utils.R
\name{get.number.of.recruits}
\alias{get.number.of.recruits}
\title{Calculates the number of (direct) recuits for each respondent.}
\usage{
get.number.of.recruits(data)
}
\arguments{
\item{data}{An rds.data.frame}
}
\description{
Calculates the number of (direct) recuits for each respondent.
}
\examples{
data(fauxmadrona)
nr <- get.number.of.recruits(fauxmadrona)
#frequency of number recruited by each id
barplot(table(nr))
}
