% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.data.frame.R
\name{get.population.size}
\alias{get.population.size}
\title{Returns the population size associated with the data.}
\usage{
get.population.size(x, check.type = TRUE)
}
\arguments{
\item{x}{the rds.data.frame}

\item{check.type}{if true, x is required to be of type rds.data.frame}
}
\description{
Returns the population size associated with the data.
}
