% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.data.frame.R
\name{get.recruitment.time}
\alias{get.recruitment.time}
\title{Returns the recruitment time for each subject}
\usage{
get.recruitment.time(
  x,
  to.numeric = TRUE,
  wave.fallback = FALSE,
  check.type = TRUE
)
}
\arguments{
\item{x}{the rds.data.frame}

\item{to.numeric}{if true, time will be converted into a numeric variable.}

\item{wave.fallback}{if true, subjects' recruitment times are ordered by wave and then by data.frame index if no recruitment time variable is available.}

\item{check.type}{if true, x is required to be of type rds.data.frame}
}
\description{
Returns the recruitment time for each subject
}
