% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gf.R
\name{gf.weights}
\alias{gf.weights}
\title{Good-Fellows weights}
\usage{
gf.weights(rds.data, outcome.variable, N = NULL, smoothed = FALSE, ...)
}
\arguments{
\item{rds.data}{An rds.data.frame}

\item{outcome.variable}{The variable used to base the weights on.}

\item{N}{Population size}

\item{smoothed}{Should the data smoothed Good-Fellows weights be computed.}

\item{...}{Unused}
}
\description{
Good-Fellows weights
}

