% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{gile.ss.weights}
\alias{gile.ss.weights}
\title{Weights using Giles SS estimator}
\usage{
gile.ss.weights(
  degs,
  N,
  number.ss.samples.per.iteration = 500,
  number.ss.iterations = 5,
  hajek = TRUE,
  SS.infinity = 0.04,
  se = FALSE,
  ...
)
}
\arguments{
\item{degs}{subjects' degrees (i.e. network sizes).}

\item{N}{Population size estimate.}

\item{number.ss.samples.per.iteration}{The number of samples to use to estimate inclusion probabilities
in a probability proportional to size without replacement design.}

\item{number.ss.iterations}{number of iterations to use in giles SS algorithm.}

\item{hajek}{Should the hajek estiamtor be used. If false, the HT estimator is used.}

\item{SS.infinity}{The sample proportion, \code{n/N}, below which the computation of the SS weights should simplify to that of the \code{RDS-II} weights.}

\item{se}{Should covariances be included.}

\item{...}{unused}
}
\description{
Weights using Giles SS estimator
}
