% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.data.frame.R
\name{has.recruitment.time}
\alias{has.recruitment.time}
\title{RDS data.frame has recruitment time information}
\usage{
has.recruitment.time(x, check.type = TRUE)
}
\arguments{
\item{x}{the rds.data.frame}

\item{check.type}{if true, x is required to be of type rds.data.frame}
}
\description{
RDS data.frame has recruitment time information
}
