% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{hcg.weights}
\alias{hcg.weights}
\title{homophily configuration graph weights}
\usage{
hcg.weights(rds.data, outcome.variable, N = NULL, small.fraction = FALSE,
  ...)
}
\arguments{
\item{rds.data}{An rds.data.frame}

\item{outcome.variable}{The variable used to base the weights on.}

\item{N}{Population size}

\item{small.fraction}{should a small sample fraction be assumed}

\item{...}{Unused}
}
\description{
homophily configuration graph weights
}
\examples{
data(fauxtime)
hcg.weights(fauxtime,"var1",N=3000)
fauxtime$NETWORK[c(1,100,40,82,77)] <- NA
}
