% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.data.frame.R
\name{is.rds.data.frame}
\alias{is.rds.data.frame}
\title{Is an instance of rds.data.frame}
\usage{
is.rds.data.frame(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\description{
Is an instance of rds.data.frame
}
