% Generated by roxygen2 (4.0.0): do not edit by hand
\name{is.rds.interval.estimate.list}
\alias{is.rds.interval.estimate.list}
\title{Is an instance of rds.interval.estimate.list
This is a (typically time ordered) sequence of RDS estimates of a comparable quantity}
\usage{
is.rds.interval.estimate.list(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\description{
Is an instance of rds.interval.estimate.list
This is a (typically time ordered) sequence of RDS estimates of a comparable quantity
}

