% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.interval.estimate.R
\name{is.rds.interval.estimate}
\alias{is.rds.interval.estimate}
\title{Is an instance of rds.interval.estimate}
\usage{
is.rds.interval.estimate(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\description{
Is an instance of rds.interval.estimate
}
