% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.R
\name{min_cluster_probability}
\alias{min_cluster_probability}
\title{min_cluster_probability}
\usage{
min_cluster_probability(k, m, l, use_gmp)
}
\arguments{
\item{k}{number of primary tumor region samples}

\item{m}{total number of metastases samples}

\item{l}{number of metastases that actually cluster together}
}
\value{
probabilility that a cluster of size l arises by chance
}
\description{
Calculates the probability that a tree where all m metastases form a cluster (have one root) exists by chance
}
