% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.R
\name{multibinomial}
\alias{multibinomial}
\title{multibinomial}
\usage{
multibinomial(m, l, cluster_numbers)
}
\arguments{
\item{m}{number of remaining cancer samples}

\item{l}{size of observed metastases cluster}

\item{cluster_numbers}{vector of possible coexisting clusters of size l for a given phylogeny}
}
\value{
product of binomials
}
\description{
Calculate the product of the binomials
}
