% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.R
\name{no_trees_min_cluster}
\alias{no_trees_min_cluster}
\title{no_trees_min_cluster}
\usage{
no_trees_min_cluster(k, m, l)
}
\arguments{
\item{k}{number of primary tumor region samples}

\item{m}{number of metastases samples}

\item{l}{minimal cluster size}
}
\value{
number of distinct tree topologies where l out of m metastases form a cluster
}
\description{
Calculates the number of trees where m metastases form a cluster of size at least l. Implements Formula (S1) from SI.
}
