% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.R
\name{no_trees_with_clusters}
\alias{no_trees_with_clusters}
\title{no_trees_with_clusters}
\usage{
no_trees_with_clusters(k, clusters)
}
\arguments{
\item{k}{number of remaining cancer samples of the same phylogeny}

\item{clusters}{vector of number of metastases samples per C clusters}
}
\value{
number of trees where the given clusters have independent origin
}
\description{
Calculates the number of trees where m1, m2, ..., mC, metastases form a cluster (have one root)
}
