% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDS.compare.proportions.R
\name{print.pvalue.table}
\alias{print.pvalue.table}
\title{Displays a pvalue.table}
\usage{
\method{print}{pvalue.table}(x, ...)
}
\arguments{
\item{x}{a pvalue.table object}

\item{...}{additional parameters passed to print.data.frame.}
}
\description{
Displays a pvalue.table
}
