% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.contingency.test.R
\name{print.rds.contin.bootstrap}
\alias{print.rds.contin.bootstrap}
\title{Displays an rds.contin.bootstrap}
\usage{
\method{print}{rds.contin.bootstrap}(x, show.table = FALSE, ...)
}
\arguments{
\item{x}{an rds.contin.bootstrap object}

\item{show.table}{Display weighted contingency table}

\item{...}{additional parameters passed to print.matrix.}
}
\description{
Displays an rds.contin.bootstrap
}
