\name{RDS-package}
\alias{RDS-package}
\alias{RDS}
\docType{package}
\title{
R Functions for Respondent-Driven Sampling.}
\description{
This package provides functions for analyzing and manipulating data collected using Respondent-Driven Sampling.
}
\details{
\tabular{ll}{
Package: \tab RDS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-04-20\cr
License: \tab GPL3\cr
LazyLoad: \tab yes\cr
}

}
\author{
W. Whipple Neely

Maintainer: W. Whipple Neely <wwneely@stat.washington.edu>
}
\references{}
\keyword{ package }
\seealso{}
\examples{}
