% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.R
\name{rds}
\alias{rds}
\title{rds}
\usage{
rds(k, m, l)
}
\arguments{
\item{k}{number of primary tumor region samples}

\item{m}{total number of metastases samples}

\item{l}{number of metastases that actually cluster together}
}
\value{
probabilility that a cluster of size l arises by chance
}
\description{
Calculates the root diversity scores as the probability that a tree where l metastases form a cluster (have one root) exists by chance as defined by Equation (2).
}
