% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_describe_pending_maintenance_actions}
\alias{rds_describe_pending_maintenance_actions}
\title{Returns a list of resources (for example, DB instances) that have at
least one pending maintenance action}
\usage{
rds_describe_pending_maintenance_actions(ResourceIdentifier, Filters,
  Marker, MaxRecords)
}
\arguments{
\item{ResourceIdentifier}{The ARN of a resource to return pending maintenance actions for.}

\item{Filters}{A filter that specifies one or more resources to return pending
maintenance actions for.

Supported filters:
\itemize{
\item \code{db-cluster-id} - Accepts DB cluster identifiers and DB cluster
Amazon Resource Names (ARNs). The results list will only include
pending maintenance actions for the DB clusters identified by these
ARNs.
\item \code{db-instance-id} - Accepts DB instance identifiers and DB instance
ARNs. The results list will only include pending maintenance actions
for the DB instances identified by these ARNs.
}}

\item{Marker}{An optional pagination token provided by a previous
\code{DescribePendingMaintenanceActions} request. If this parameter is
specified, the response includes only records beyond the marker, up to a
number of records specified by \code{MaxRecords}.}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that the remaining
results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}
}
\description{
Returns a list of resources (for example, DB instances) that have at
least one pending maintenance action.
}
\section{Request syntax}{
\preformatted{rds$describe_pending_maintenance_actions(
  ResourceIdentifier = "string",
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  Marker = "string",
  MaxRecords = 123
)
}
}

\keyword{internal}
