   \name{simulated}
          \docType{data}
          \alias{simulated.data}
          \title{A Simulated RDS Data Set}
          \description{
          This is a simulated data set used to demonstrate RDS functions and analysis.
          }
          \usage{simulated}
          \format{An R data frame with 389 observations.}
          \source{}
          \references{
          }
          \keyword{datasets}
