\name{transition.counts.to.Markov.mle}
\alias{transition.counts.to.Markov.mle}
\title{calculates the mle. ie.e the row proportions of the transition matrix}
\usage{
  transition.counts.to.Markov.mle(transition.counts)
}
\arguments{
  \item{transition.counts}{a matrix or table of transition
  counts}
}
\description{
  calculates the mle. ie.e the row proportions of the
  transition matrix
}
\details{
  depricated. just useprop.table(transition.counts,1)
}

