% Generated by roxygen2 (4.0.0): do not edit by hand
\name{write.gephi}
\alias{write.gephi}
\title{Writes out the RDS tree in Gephi format}
\usage{
write.gephi(x, file = NULL, scale = 20)
}
\arguments{
\item{x}{An rds.data.frame.}

\item{file}{a character vector representing a file.}

\item{scale}{a positive scaler to scale the outputed positions.}
}
\description{
Writes out the RDS tree in Gephi format
}
\details{
All the variables in the RDS data set are exported as well.
\code{paste0(file,".GEXF")} contains the nodal attributes and edge information.
}

