\name{write.graphviz}
\alias{write.graphviz}
\title{writes an rds.data.frame recruitment tree as a GraphViz file}
\usage{
  write.graphviz(x, file)
}
\arguments{
  \item{x}{An rds.data.frame.}

  \item{file}{A character vector representing the file}
}
\description{
  writes an rds.data.frame recruitment tree as a GraphViz
  file
}

