% Generated by roxygen2 (4.0.0): do not edit by hand
\name{write.rdsat}
\alias{write.rdsat}
\title{Writes out the RDS tree in RDSAT format}
\usage{
write.rdsat(x, file = NULL)
}
\arguments{
\item{x}{An rds.data.frame.}

\item{file}{a character vector representing a file.}
}
\description{
Writes out the RDS tree in RDSAT format
}

