% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-export.R
\name{write.rdsobj}
\alias{write.rdsobj}
\title{Export an rds.data.frame to file}
\usage{
write.rdsobj(x, file)
}
\arguments{
\item{x}{The rds.data.frame to export}

\item{file}{The name of the file to create.}
}
\description{
Export an rds.data.frame to file
}
