% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.char}
\alias{as.char}
\title{converts to character with minimal loss of precision for numeric variables}
\usage{
as.char(x, ...)
}
\arguments{
\item{x}{the value}

\item{...}{passed to either format or as.character.}
}
\description{
converts to character with minimal loss of precision for numeric variables
}
