% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.data.frame.R
\name{assert.valid.rds.data.frame}
\alias{assert.valid.rds.data.frame}
\title{Does various checks and throws errors if x is not a valid rds.data.frame}
\usage{
assert.valid.rds.data.frame(x, ...)
}
\arguments{
\item{x}{an rds.data.frame}

\item{...}{unused}
}
\description{
Does various checks and throws errors if x is not a valid rds.data.frame
}
\details{
Throws an informative message if x is malformed.
}
