% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-plots.R
\name{cumulative.estimate}
\alias{cumulative.estimate}
\title{Calculates estimates at each successive wave of the sampling process}
\usage{
cumulative.estimate(
  rds.data,
  outcome.variable,
  est.func = RDS.II.estimates,
  n.eval.points = 25,
  ...
)
}
\arguments{
\item{rds.data}{An rds.data.frame}

\item{outcome.variable}{The outcome}

\item{est.func}{A function taking rds.data and outcome.variable as parameters and
returning an rds.weighted.estimate object}

\item{n.eval.points}{number of evaluation points to calculate the estimates at}

\item{...}{additional parameters for est.func}
}
\description{
Calculates estimates at each successive wave of the sampling process
}
