% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.data.frame.R
\docType{methods}
\name{[<-.rds.data.frame}
\alias{[<-.rds.data.frame}
\alias{[<-,rds.data.frame-method}
\title{indexing}
\usage{
\method{[}{rds.data.frame} (x, i, j) <- value
}
\arguments{
\item{x}{object}

\item{i}{indices}

\item{j}{indices}

\item{value}{value}
}
\description{
indexing
}
\details{
Indexed assignment. If the result is not a valid rds.data.frame, an error
is emitted.
}
