% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.data.frame.R
\name{get.id}
\alias{get.id}
\title{Get the subject id}
\usage{
get.id(x, check.type = TRUE)
}
\arguments{
\item{x}{an rds.data.frame object}

\item{check.type}{if true, x is required to be of type rds.data.frame}
}
\description{
Get the subject id
}
\details{
returns the variable indicated by the 'id' attribute, coercing to
a character vector
}
