% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rds.data.frame.R
\name{get.rid}
\alias{get.rid}
\title{Get recruiter id}
\usage{
get.rid(x, check.type = TRUE)
}
\arguments{
\item{x}{an rds.data.frame object}

\item{check.type}{if true, x is required to be of type rds.data.frame}
}
\description{
Get recruiter id
}
\details{
returns the variable indicated by the 'recruiter.id' attribute, coersing to
a character vector
}

