% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.R
\name{klm}
\alias{klm}
\title{klm}
\usage{
klm(phy)
}
\arguments{
\item{phy}{Either a phylo-object (as from ape::nj()) or a distance matrix (square matrix).}
}
\value{
data.frame where rows are metastasis sample-types and columns are k, l, m, and rds values.
}
\description{
klm
}
