% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.data.frame.R
\name{print.rds.data.frame}
\alias{print.rds.data.frame}
\title{Displays an rds.data.frame}
\usage{
\method{print}{rds.data.frame}(x, ...)
}
\arguments{
\item{x}{an rds.data.frame object}

\item{...}{additional parameters passed to print.data.frame.}
}
\description{
Displays an rds.data.frame
}
