% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.rds.interval.estimate}
\alias{print.rds.interval.estimate}
\title{Prints an \code{rds.interval.estimate} object}
\usage{
\method{print}{rds.interval.estimate}(x, ...)
}
\arguments{
\item{x}{an \code{rds.interval.estimate} object}

\item{...}{unused}
}
\description{
Prints an \code{rds.interval.estimate} object
}

