% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDS-package.R
\docType{package}
\name{RDS-package}
\alias{RDS}
\alias{RDS-package}
\title{RDS: Respondent-Driven Sampling}
\description{
Provides functionality for carrying out estimation with data collected using Respondent-Driven Sampling. This includes Heckathorn's RDS-I and RDS-II estimators as well as Gile's Sequential Sampling estimator. The package is part of the "RDS Analyst" suite of packages for the analysis of respondent-driven sampling data. See Gile and Handcock (2010) \doi{10.1111/j.1467-9531.2010.01223.x}, Gile and Handcock (2015) \doi{10.1111/rssa.12091} and Gile, Beaudry, Handcock and Ott (2018) \doi{10.1146/annurev-statistics-031017-100704}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hpmrg.org}
}

}
\author{
\strong{Maintainer}: Mark S. Handcock \email{handcock@stat.ucla.edu} (\href{https://orcid.org/0000-0002-9985-2785}{ORCID})

Authors:
\itemize{
  \item Krista J. Gile \email{gile@math.umass.edu}
  \item Ian E. Fellows \email{ian@fellstat.com}
}

Other contributors:
\itemize{
  \item W. Whipple Neely \email{wwneely@stat.washington.edu} [contributor]
}

}
\keyword{internal}
