% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rds-doc.R
\docType{package}
\name{RDS}
\alias{RDS}
\alias{RDS-package}
\title{This package provides functionality for carrying out estimation
    with data collected using Respondent-Driven Sampling. This includes
    Heckathorn's RDS-I and RDS-II estimators as well as Gile's Sequential
    Sampler estimator.}
\description{
This package provides functionality for carrying out estimation
    with data collected using Respondent-Driven Sampling. This includes
    Heckathorn's RDS-I and RDS-II estimators as well as Gile's Sequential
    Sampler estimator.
}

