% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.data.frame.R
\name{show.rds.data.frame}
\alias{show.rds.data.frame}
\title{Displays an rds.data.frame}
\usage{
show.rds.data.frame(x, ...)
}
\arguments{
\item{x}{an rds.data.frame object.}

\item{...}{additional parameters passed to print.data.frame.}
}
\description{
Displays an rds.data.frame
}
