% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-utils.R
\name{transition.counts.to.Markov.mle}
\alias{transition.counts.to.Markov.mle}
\title{calculates the mle. i.e. the row proportions of the transition matrix}
\usage{
transition.counts.to.Markov.mle(transition.counts)
}
\arguments{
\item{transition.counts}{a matrix or table of transition counts}
}
\description{
calculates the mle. i.e. the row proportions of the transition matrix
}
\details{
depreicated. just use prop.table(transition.counts,1)
}
